==> KDD/3_KDD_merge_each.ipynb <==
import pandas as pd 
import numpy as np  
from sklearn.utils import shuffle

dt11 = pd.read_csv("../dataset/KDD-2018/trainprobe.csv")
dt12 = pd.read_csv("../dataset/KDD-2018/generated_probe.csv")
dt21 = pd.read_csv("../dataset/KDD-2018/trainr2l.csv")
dt22 = pd.read_csv("../dataset/KDD-2018/generated_r2l.csv")

print(dt11['Label'].value_counts())
print(dt21['Label'].value_counts())

a1 = dt12.sample(n=8366, replace=False, random_state=1)
a2 = dt22.sample(n=13513, replace=False, random_state=1)

# Merge data
data1 = pd.concat([dt11,a1],ignore_index = True,axis=0)
data_balanced1 = shuffle(data1).reset_index(drop=True)
data_balanced1.to_csv("../dataset/KDD-2018/balanced_probe.csv", index=False, header = True)

data2 = pd.concat([dt21,a2],ignore_index = True,axis=0)
data_balanced2 = shuffle(data2).reset_index(drop=True)
data_balanced2.to_csv("../dataset/KDD-2018/balanced_r2l.csv", index=False, header = True)


print("Balanced dataset:")
print(data_balanced1['Label'].value_counts())
print("-----------------------------------------")
print(data_balanced2['Label'].value_counts())
print("-----------------------------------------")
