==> KDD/4_merge_full.ipynb <==
# Read original data
import numpy as np
import pandas as pd
from sklearn.utils import shuffle

dt = pd.read_csv("../dataset/KDD-2018/train_ori.csv")
dt1 = pd.read_csv("../dataset/KDD-2018/generated_probe.csv")
dt2 = pd.read_csv("../dataset/KDD-2018/generated_r2l.csv")
dt3 = pd.read_csv("../dataset/KDD-2018/generated_u2r.csv")

# Convert label
print("Label converting...")
label = {1: 'Probe'}
dt1['Label'] = dt1['Label'].map(label).fillna(dt1['Label'])

label2 = {1: 'R2L'}
dt2['Label'] = dt2['Label'].map(label2).fillna(dt2['Label'])

label3 = {1: 'U2R'}
dt3['Label'] = dt3['Label'].map(label3).fillna(dt3['Label'])

print('Completed!')

print(dt['Label'].value_counts())
print(dt1['Label'].value_counts())
print(dt2['Label'].value_counts())

a1 = dt1.sample(n=8366, replace=False, random_state=1)
a2 = dt2.sample(n=13513, replace=False, random_state=1)
a3 = dt3.sample(n=13995, replace=False, random_state=1)

# Merge data
data = pd.concat([dt,a1,a2,a3],ignore_index = True,axis=0)
data_balanced = shuffle(data).reset_index(drop=True)
data_balanced.to_csv("../dataset/KDD-2018/train.csv", index=False, header = True)
print("Original dataset:")
print(dt['Label'].value_counts())
print("-----------------------------------------")
print("Balanced dataset:")
print(data_balanced['Label'].value_counts())
print("-----------------------------------------")
